Using CULong for CSocket and proper typing

+++ Network/Socket/Handle.hs
@@ -21,9 +21,10 @@ import Network.Socket.Types
 -- sequence with appropriate handshakes specified by the protocol.
 
 -- TODO: WinIO doesn't use fd, add support
+--       Need to remove fromIntegral.
 socketToHandle :: Socket -> IOMode -> IO Handle
 socketToHandle s mode = invalidateSocket s err $ \oldfd -> do
-    h <- fdToHandle' oldfd (Just GHC.IO.Device.Stream) True (show s) mode True{-bin-}
+    h <- fdToHandle' (fromIntegral oldfd) (Just GHC.IO.Device.Stream) True (show s) mode True{-bin-}
     hSetBuffering h NoBuffering
     return h
   where
